Create attempt_three.py
starting a new attempt and taking it a lot slower and making sure I understand all the code. So far I login to my email and retrieve the email ids. I then get the last email id (the newest email in inbox) and fetch the data from it. I then get the email address and the subject of the email